+++ src/pages/Archive.tsx
@@ -1,0 +1,5 @@
+function ArchivesPage() {
+  return <div>Test</div>;
+}
+
+export default ArchivesPage;
